Move JSON examples to a more suitable package
